# New Span-based serialization logic (followup for #7351)