… csharp/src/Google.Protobuf.Benchmarks/WrapperBenchmarkMessages.cs
… csharp/src/Google.Protobuf.Benchmarks/WriteMessagesBenchmark.cs
… csharp/src/Google.Protobuf.Benchmarks/WriteRawPrimitivesBenchmark.cs
… csharp/src/Google.Protobuf.Conformance/Conformance.cs
… csharp/src/Google.Protobuf.Test.TestProtos/MapUnittestProto3.cs
… csharp/src/Google.Protobuf.Test.TestProtos/TestMessagesProto2.cs
… csharp/src/Google.Protobuf.Test.TestProtos/TestMessagesProto3.cs
… csharp/src/Google.Protobuf.Test.TestProtos/Unittest.cs
… csharp/src/Google.Protobuf.Test.TestProtos/UnittestCustomOptionsProto3.cs
… csharp/src/Google.Protobuf.Test.TestProtos/UnittestImport.cs
… csharp/src/Google.Protobuf.Test.TestProtos/UnittestImportProto3.cs
… csharp/src/Google.Protobuf.Test.TestProtos/UnittestImportPublic.cs
… csharp/src/Google.Protobuf.Test.TestProtos/UnittestImportPublicProto3.cs
… csharp/src/Google.Protobuf.Test.TestProtos/UnittestIssue6936B.cs
… csharp/src/Google.Protobuf.Test.TestProtos/UnittestIssue6936C.cs
… csharp/src/Google.Protobuf.Test.TestProtos/UnittestIssues.cs
… csharp/src/Google.Protobuf.Test.TestProtos/UnittestProto3.cs
… csharp/src/Google.Protobuf.Test.TestProtos/UnittestProto3Optional.cs
… csharp/src/Google.Protobuf.Test.TestProtos/UnittestSelfreferentialOptions.cs
… csharp/src/Google.Protobuf.Test.TestProtos/UnittestWellKnownTypes.cs
… csharp/src/Google.Protobuf.Test/Buffers/ArrayBufferWriter.cs
… csharp/src/Google.Protobuf.Test/ByteStringTest.cs
… csharp/src/Google.Protobuf.Test/CodedOutputStreamTest.cs
… csharp/src/Google.Protobuf.Test/ExtensionSetTest.cs
… csharp/src/Google.Protobuf.Test/FieldCodecTest.cs
… csharp/src/Google.Protobuf.Test/GeneratedMessageTest.Proto2.cs
… csharp/src/Google.Protobuf.Test/GeneratedMessageTest.cs
… csharp/src/Google.Protobuf.Test/LegacyGeneratedCodeTest.cs
… csharp/src/Google.Protobuf.Test/MessageParsingHelpers.cs
… csharp/src/Google.Protobuf.Test/WellKnownTypes/WrappersTest.cs