Span-based parsing has been added by https://github.com/protocolbuffers/protobuf/pull/7351 , this PR adds the serialization counterpart, which allows serialization to `IBufferWriter` (which is much more GC friendly).

Based on https://github.com/protocolbuffers/protobuf/pull/7490

Update 2020-06-29:
The new benchmarks (as of commit 4f0afc7) are here: https://gist.github.com/jtattermusch/c9e5389a398f7c79710a9bf12aee215f
For both writing primitives and writing messages the performance is comparable (there are some significant improvement and some slight regressions) to the original state, which seems good enough (as this PR offers an API to serialize directly into a Span and/or to IBufferWriter, which are modern APIs allowing for much better GC efficiency).