@JamesNK 2020-06-02
Is `IBufferWriter<byte>` supported yet? It is referenced but at the moment all tests write to `byte[]` or `Stream`.