@jtattermusch 2020-06-02
> Is `IBufferWriter<byte>` supported yet? It is referenced but at the moment all tests write to `byte[]` or `Stream`.

I still need to expose the API and add tests, but vast majority of the internal implementation is already there.