@jtattermusch 2020-06-03
I added two MessageExtensions:
`WriteTo(this IMessage message, IBufferWriter<byte> output)` and `WriteTo(this IMessage message, Span<byte> output)`. 

Also added test coverage for the WriteContext based serialization.

I still need to add some benchmarks.