@jtattermusch 2020-06-08
I addressed the comments, resolved several TODOs and I think this is very close to being mergeable.
I've also run most of the benchmarks and compared before and after performance and so far things are looking pretty good (there are basically no regressions and there are some improvements) - I will publish the full comparison later.