@jtattermusch 2020-06-08
@haberman  this is now ready for review. @JamesNK will review the code from C# perspective, we just need a high-level protobuf review. Most of the ideas are analogous to what we've done in #7351, but this time it's a bit simpler (serialization is easier than parsing).