@JamesNK 2020-06-09
It would be useful to have the equivalent of ParseMessagesBenchmark but for writing. That would show the end-to-end perf on complex objects.