@jtattermusch 2020-06-11
Before and after benchmark results is here:
https://gist.github.com/jtattermusch/2d8cc4eab54a53de1079132495d3f55d

Summary 
- for WriteRawPrimitivesBenchmark, there are some primitives that are slighly slower, but most of them are faster and some of the much are much faster.
- for WriteMessageBenchmark, unfortunately it seems that writing of messages is actually quite a bit slower, so that's an area that still needs to be investigated (not sure what the problem is).