@jtattermusch 2020-06-11
@JamesNK  can you please take a look at the WriteMessagesBenchmark results from my comment https://github.com/protocolbuffers/protobuf/pull/7576#issuecomment-642698060 and see if:
- you're getting the same results
- there are some optimizations I've missed and that would speed up WriteContext-based writing to an acceptable level.