@JamesNK 2020-06-11
Hmmm, interesting. In my PR it was writing that ended up faster, and parsing was the same speed.

My WriteMessagesBenchmark results: https://gist.github.com/JamesNK/86d8707398326904dfbccd3744022960

One optimization is range checks - https://github.com/protocolbuffers/protobuf/pull/7576#discussion_r437119673. I don't think that explains the difference. My guess is it is probably something to do with copying structs.