@jtattermusch 2020-06-12
@JamesNK  thanks for the results. It seems that on windows the ManyWrapperFields and Empty serialization has ~similar speed as before, but there's a big regression in terms of serializing ManyPrimitiveFieldsMessage.
On Linux the results have a similar trend but it isn't as clear-cut.

I guess it might have to do with passing the `ref buffer, ref state` to each of the writing primitive method (and each of the writing operations is very fast as it doesn't really do much) and the overhead builds up.