@jtattermusch 2020-06-12
It turns out that the comparably bad performance with ManyPrimitiveFields message is mostly caused by using AggresiveInlining heavily for all the primitives. What I think happens is:
- the ManyPrimitiveFieldsMessage.InternalWriteTo method has a lot of invocations to primitives, and they all get inlined.
- the InternalWriteToMethod becomes too complex, which prevents the JIT from noticing the fact that most of the message fields are unused and only the few existing ones are going to be used. 
- as a result, checking whether the message fields are set takes much longer, which results in much slower serialization (where most time is spent checking whether the fields are set or not).

The solution is removing some of the AggresiveInlining annotations, but on the other hand that needs to be done carefully, because if too many are removed, the WriteRawPrimitivesBenchmarks results deteriorate.
So I think the resolution is finding the right balance between which AggresiveInlining annotations to keep and which ones to remove. Based on my experiments, we might not be able to reach 100% of the original performance in all cases unfortunately (but altogether this PR probably still worth accepting because it allows GC friendly serialization directly to the IBufferWriter with no additional copies).