@JamesNK 2020-06-12
Good find.

I think the best improvement here is reducing the amount of code in the primitive methods, and the best way to do that is optimize for the fast path of there being enough data in the buffer.

See my comment here about removing the WriteRawByte calls and just having a single call to refresh the buffer if there is not enough room - https://github.com/protocolbuffers/protobuf/pull/7576#discussion_r437118669

What `CodedOutputWriter` does:
```cs
        internal void WriteRawLittleEndian32(uint value)
        {
            const int length = 4;

            Ensure(length);

            BinaryPrimitives.WriteUInt32LittleEndian(span.Slice(buffered), value);

            buffered += length;
        }
```

What WritingPrimitives.cs does today:
```cs
        [MethodImpl(MethodImplOptions.AggressiveInlining)]
        public static void WriteRawLittleEndian32(ref Span<byte> buffer, ref WriterInternalState state, uint value)
        {
            const int length = sizeof(uint);
            if (state.position + length > state.limit)
            {
                WriteRawByte(ref buffer, ref state, (byte)value);
                WriteRawByte(ref buffer, ref state, (byte)(value >> 8));
                WriteRawByte(ref buffer, ref state, (byte)(value >> 16));
                WriteRawByte(ref buffer, ref state, (byte)(value >> 24));
            }
            else
            {
                BinaryPrimitives.WriteUInt32LittleEndian(buffer.Slice(state.position), value);
                state.position += length;
            }
        }
```

Improving WritingPrimitives.cs:
```cs
        [MethodImpl(MethodImplOptions.AggressiveInlining)]
        public static void WriteRawLittleEndian32(ref Span<byte> buffer, ref WriterInternalState state, uint value)
        {
            const int length = sizeof(uint);
            if (state.position + length > state.limit)
            {
                // Rarely called. RefreshBuffer should have NoInlining on it.
                RefreshBuffer(ref buffer, ref state, requireLength: length);
            }
            
            BinaryPrimitives.WriteUInt32LittleEndian(buffer.Slice(state.position), value);
            state.position += length;
        }
```

You don't need to worry about using up the entire buffer. It shouldn't matter to the Stream or IBufferWriter that sometimes you are writing/advancing 4096 bytes, and other times writing/advancing 4091 bytes.