@JamesNK 2020-06-12
Useful code from CodedOutputWriter - https://github.com/JamesNK/protobuf/blob/155774996be69135c05b30960e6ef016c530894c/csharp/src/Google.Protobuf/CodedOutputWriter.cs#L626-L655