@JamesNK 2020-06-22
Optimize writing strings across buffers: https://github.com/JamesNK/protobuf/commit/99e830bf6f020da98620c46beffa052295c08504

Before:
```
|                                               Method | BytesToWrite | encodedSize |       Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------------------------- |------------- |------------ |-----------:|----------:|----------:|-------:|------:|------:|----------:|
| WriteString_WriteContextBufferWriter_LimitBufferSize |        10080 |           1 | 150.550 us | 1.6307 us | 1.5254 us |      - |     - |     - |       2 B |
| WriteString_WriteContextBufferWriter_LimitBufferSize |        10080 |           4 |  49.011 us | 0.6282 us | 0.5569 us |      - |     - |     - |         - |
| WriteString_WriteContextBufferWriter_LimitBufferSize |        10080 |          10 |  26.207 us | 0.4905 us | 0.4348 us | 0.0610 |     - |     - |     320 B |
| WriteString_WriteContextBufferWriter_LimitBufferSize |        10080 |         105 |  11.951 us | 0.0635 us | 0.0530 us | 0.2747 |     - |     - |    1152 B |
| WriteString_WriteContextBufferWriter_LimitBufferSize |        10080 |       10080 |   3.178 us | 0.0628 us | 0.0616 us | 2.4071 |     - |     - |   10104 B |
```
After:
```
|                                               Method | BytesToWrite | encodedSize |       Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------------------------- |------------- |------------ |-----------:|----------:|----------:|-------:|------:|------:|----------:|
| WriteString_WriteContextBufferWriter_LimitBufferSize |        10080 |           1 | 161.463 us | 1.2570 us | 1.1758 us |      - |     - |     - |         - |
| WriteString_WriteContextBufferWriter_LimitBufferSize |        10080 |           4 |  49.170 us | 0.7292 us | 0.9223 us |      - |     - |     - |         - |
| WriteString_WriteContextBufferWriter_LimitBufferSize |        10080 |          10 |  23.225 us | 0.1782 us | 0.1667 us |      - |     - |     - |         - |
| WriteString_WriteContextBufferWriter_LimitBufferSize |        10080 |         105 |   9.763 us | 0.0765 us | 0.0679 us |      - |     - |     - |         - |
| WriteString_WriteContextBufferWriter_LimitBufferSize |        10080 |       10080 |   2.658 us | 0.0195 us | 0.0173 us | 0.0381 |     - |     - |     160 B |
```